Add state and onChange methods to components
Each component, except for the Skills component, now holds properties in
state, that are updated when the user types in the appropriate field.
The reason why I didn't yet do this for the Skills component, is because
I want to design it differently. I want the user to be able to type in
a skill, click the enter key, then have that skill appear as an element
above the skill label.