Add delete skill functionality
Additionally, because of the nature of having to add and remove skill
elements and data, I decided to move the skills properties in state to
the App parent component state. Now, it is the App component that
handles the state of the skills component. So now, when a skill is
added in the form, it is added in state in the App component, and an
element with the name of the added skill is created in the app component
and displayed in the UI. Finally, when the skill element's x icon is
clicked, both the element and property are removed from the App
component's state.