Style structure of the app
This is the main styling regarding the structure of the app. Most
elements are placed exactly where I want them to be. All I need to worry
about are a couple of details and the coloring of the app, and the
styling of the form will be complete. Regarding the styling of this
project, I decided to use a main style sheet for the main elements and
styling in the App component, and individual style modules for each
component.