Apply a color palette to app
Besides choosing and implementing a color palette, I declared css
variables in each css module, to represent all the colors found in the
app. I feel like there is more I can do to the color palette, to make
the app pop more, but due to time constraints, we'll just leave it like
that.