Create logic for passing state props up to App
I had to rework some logic in order for the parent App component to gain
access to the state properties of all of the children components, minus
the Skills component. This mainly involved Creating a callback function
method in the App component to be passed as a prop to the children
components. This callback function is passed via the onDataChange
attribute, which is very helpful, because the data is updated and sent
up as soon as the user starts adding to a field. I also added some logic
in the remove methods of the App component so that the data from the
last instance of the component gets deleted. This is able to be done
because the data of a component instance is given an index and placed
in the appropriate array in state in the App component. Now I can start
generating the pdf.