Create PDF document using state properties
I decided to use a package called react-pdf renderer in order to
generate the pdf. This was not easy to learn how to use however, and my
pdf does not generate exactly the way I want it to. I was having no luck
styling elements using flex-box. Originally, I wanted the dates to be on
the same line as work, project, and education titles, and to space the
elements using justify content, but I did not have any luck doing this.
Because this took longer that I thought it would, I am going to leave
this project the way it is for now.