Convert class components to function components
There are still two errors that need fixing:

1. In the general information component; when the user starts filling
out the name field, and auto completes the rest, some of that
information does not show up on the final cv.

2. For the work, projects, and education sections; as the user types in
the fields, each key press saves into state, but does not properly
override what is necessary. This means that in the final cv, each
individual key press is a snapshot in time that is saved in state and is
shown in the cv.